+++ fn/dom-collections/iterator.js
@@ -1,2 +1,2 @@
-require('../../modules/web.dom.iterable');
+require('../../modules/web.dom-collections.iterator');
 module.exports = require('../../modules/_core').Array.values;