+++ web/dom-collections.js
@@ -1,3 +1,3 @@
-require('../modules/web.dom.for-each');
-require('../modules/web.dom.iterable');
+require('../modules/web.dom-collections.for-each');
+require('../modules/web.dom-collections.iterator');
 module.exports = require('../modules/_core');